# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sogram social trading platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TraderTier = (typeof TraderTier)[keyof typeof TraderTier];

export const TraderTier = {
  Elite: "Elite",
  Expert: "Expert",
  Rising: "Rising",
} as const;

export interface Trader {
  id: number;
  username: string;
  handle: string;
  bio?: string;
  repScore: number;
  tier: TraderTier;
  isVerified: boolean;
  followersCount: number;
  copiersCount: number;
  signalsCount: number;
  pnl30d: string;
  winRate: number;
  avgRR?: number;
  maxDrawdown?: number;
  avgHoldDays?: number;
  sharpeRatio?: number;
  joinedAt: string;
}

export type TradeSide = (typeof TradeSide)[keyof typeof TradeSide];

export const TradeSide = {
  LONG: "LONG",
  SHORT: "SHORT",
} as const;

export interface Trade {
  id: number;
  traderId: number;
  traderUsername: string;
  traderHandle: string;
  traderRepScore: number;
  traderTier: string;
  asset: string;
  side: TradeSide;
  entryPrice: number;
  exitPrice: number;
  pnlUsd: string;
  pnlPct: string;
  positionSize: string;
  leverage: string;
  duration: string;
  caption: string;
  tags: string[];
  isOnChainVerified: boolean;
  likes: number;
  copies: number;
  comments: number;
  createdAt: string;
}

export type TraderProfile = Trader & {
  recentTrades: Trade[];
  pnlCurve: number[];
};

export interface CreateTraderBody {
  username: string;
  handle: string;
  bio?: string;
  walletAddress?: string;
}

export type CreateTradeBodySide =
  (typeof CreateTradeBodySide)[keyof typeof CreateTradeBodySide];

export const CreateTradeBodySide = {
  LONG: "LONG",
  SHORT: "SHORT",
} as const;

export interface CreateTradeBody {
  traderId: number;
  asset: string;
  side: CreateTradeBodySide;
  entryPrice: number;
  exitPrice: number;
  pnlUsd: string;
  pnlPct: string;
  positionSize: string;
  leverage: string;
  duration: string;
  caption: string;
  tags?: string[];
  isOnChainVerified?: boolean;
}

export type SignalSide = (typeof SignalSide)[keyof typeof SignalSide];

export const SignalSide = {
  LONG: "LONG",
  SHORT: "SHORT",
} as const;

export interface Signal {
  id: number;
  traderId: number;
  traderUsername: string;
  traderHandle: string;
  traderRepScore: number;
  traderTier: string;
  asset: string;
  side: SignalSide;
  entryZone: string;
  target: string;
  stopLoss: string;
  rrRatio: string;
  confidence: number;
  caption: string;
  tags: string[];
  likes: number;
  followers: number;
  comments: number;
  createdAt: string;
}

export type CreateSignalBodySide =
  (typeof CreateSignalBodySide)[keyof typeof CreateSignalBodySide];

export const CreateSignalBodySide = {
  LONG: "LONG",
  SHORT: "SHORT",
} as const;

export interface CreateSignalBody {
  traderId: number;
  asset: string;
  side: CreateSignalBodySide;
  entryZone: string;
  target: string;
  stopLoss: string;
  rrRatio: string;
  confidence: number;
  caption: string;
  tags?: string[];
}

export type FeedItemType = (typeof FeedItemType)[keyof typeof FeedItemType];

export const FeedItemType = {
  trade: "trade",
  signal: "signal",
} as const;

export interface FeedItem {
  type: FeedItemType;
  trade?: Trade;
  signal?: Signal;
}

export interface FeedResponse {
  items: FeedItem[];
  total: number;
}

export interface CopyConfig {
  id: number;
  copierId: number;
  leaderId: number;
  leaderUsername: string;
  leaderHandle: string;
  leaderRepScore: number;
  leaderPnl30d: string;
  leaderWinRate: number;
  copyRatioPct: number;
  maxPerTradeUsd: number;
  stopCopyDrawdownPct: number;
  allowedPairs: string[];
  copyPerps: boolean;
  copySpot: boolean;
  isActive: boolean;
  todayPnl?: string;
  openPositions?: number;
  deployedUsd?: number;
}

export interface UpsertCopyConfigBody {
  copierId: number;
  leaderId: number;
  copyRatioPct: number;
  maxPerTradeUsd: number;
  stopCopyDrawdownPct: number;
  allowedPairs: string[];
  copyPerps: boolean;
  copySpot: boolean;
  isActive: boolean;
}

export interface PairAnalytics {
  pair: string;
  shareOfTraders: number;
  volume: string;
  longsPct: number;
  shortsPct: number;
}

export interface LeverageBucket {
  label: string;
  pct: number;
  count: number;
}

export type MacroEventDirection =
  (typeof MacroEventDirection)[keyof typeof MacroEventDirection];

export const MacroEventDirection = {
  up: "up",
  down: "down",
  neutral: "neutral",
} as const;

export interface MacroEvent {
  label: string;
  value: string;
  note: string;
  direction: MacroEventDirection;
}

export interface AnalyticsSummary {
  crowdPositions: PairAnalytics[];
  leverageDistribution: LeverageBucket[];
  overallLongPct: number;
  overallShortPct: number;
  avgLeverage: number;
  activeTraders: number;
  macroEvents: MacroEvent[];
}

export interface MarketTicker {
  symbol: string;
  price: string;
  change24h: string;
  changePct24h: string;
  isUp: boolean;
}

export interface MarketPrices {
  tickers: MarketTicker[];
  updatedAt: string;
}

export type WhalePositionSide =
  (typeof WhalePositionSide)[keyof typeof WhalePositionSide];

export const WhalePositionSide = {
  LONG: "LONG",
  SHORT: "SHORT",
} as const;

export interface WhalePosition {
  traderUsername: string;
  traderHandle: string;
  repScore: number;
  positionSizeUsd: string;
  pair: string;
  side: WhalePositionSide;
  leverage: string;
  timeAgo: string;
  traderId: number;
}

export interface LeaderboardEntry {
  rank: number;
  id: number;
  username: string;
  handle: string;
  repScore: number;
  tier: string;
  pnl30d: string;
  winRate: number;
  copiersCount: number;
  totalTrades: number;
}

export type GetFeedParams = {
  filter?: GetFeedFilter;
  limit?: number;
  offset?: number;
};

export type GetFeedFilter = (typeof GetFeedFilter)[keyof typeof GetFeedFilter];

export const GetFeedFilter = {
  all: "all",
  wins: "wins",
  signals: "signals",
} as const;

export type ListTradersParams = {
  limit?: number;
};

export type ListTraders200 = {
  traders: Trader[];
};

export type GetTraderTradesParams = {
  limit?: number;
};

export type GetTraderTrades200 = {
  trades: Trade[];
};

export type LikeTrade200 = {
  likes: number;
  liked: boolean;
};

export type LikeSignal200 = {
  likes: number;
  liked: boolean;
};

export type ListCopyConfigsParams = {
  copierId: number;
};

export type ListCopyConfigs200 = {
  configs: CopyConfig[];
};

export type GetWhaleActivity200 = {
  whales: WhalePosition[];
};

export type GetLeaderboardParams = {
  limit?: number;
};

export type GetLeaderboard200 = {
  traders: LeaderboardEntry[];
};